% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomChisquaredArray-class.R
\docType{class}
\name{RandomChisquaredArray-class}
\alias{RandomChisquaredArray-class}
\alias{RandomChisquaredArraySeed-class}
\alias{sampleDistrParam,RandomChisquaredArraySeed-method}
\alias{sampleDistrFun,RandomChisquaredArraySeed-method}
\alias{matrixClass,RandomChisquaredArraySeed-method}
\alias{RandomChisquaredArraySeed}
\alias{DelayedArray,RandomChisquaredArraySeed-method}
\alias{RandomChisquaredArray}
\title{DelayedArray of chunked chi-squared-distributed values}
\usage{
RandomChisquaredArraySeed(dim, df, ncp = 0, chunkdim = NULL)

\S4method{DelayedArray}{RandomChisquaredArraySeed}(seed)

RandomChisquaredArray(dim, df, ncp = 0, chunkdim = NULL)
}
\arguments{
\item{dim}{Integer vector of positive length, specifying the dimensions of the array.}

\item{df, ncp}{Numeric vector used as the argument of the same name in \code{\link{qchisq}}.
Alternatively, a numeric array-like object with the same dimensions as \code{dim}.}

\item{chunkdim}{Integer vector of length equal to \code{dim}, containing the dimensions of each chunk.}

\item{seed}{A RandomChisquaredArraySeed object.}
}
\value{
All constructors return an instance of a RandomChisquaredArray object,
containing random draws from a chi-squared distribution with the specified parameters.
}
\description{
A \linkS4class{DelayedArray} subclass that performs on-the-fly sampling of chi-squared-distributed values.
}
\examples{
X <- RandomChisquaredArraySeed(c(1e5, 1e5), df=5)
Y <- DelayedArray(X)
Y

# Fiddling with the distribution parameters:
X2 <- RandomChisquaredArraySeed(c(1e5, 1e5), df=runif(1e5)*20)
Y2 <- DelayedArray(X2)
Y2

# Using another array as input:
library(Matrix)
df <- rsparsematrix(1e5, 1e5, density=0.00001)
df <- abs(DelayedArray(df) + 1) * 10
X3 <- RandomChisquaredArraySeed(c(1e5, 1e5), df=df)
Y3 <- DelayedArray(X3)
Y3

}
\seealso{
The \linkS4class{RandomArraySeed} class, for details on chunking and the distributional parameters.
}
\author{
Aaron Lun
}
