% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomExponentialArray-class.R
\docType{class}
\name{RandomExponentialArray-class}
\alias{RandomExponentialArray-class}
\alias{RandomExponentialArraySeed-class}
\alias{sampleDistrParam,RandomExponentialArraySeed-method}
\alias{sampleDistrFun,RandomExponentialArraySeed-method}
\alias{matrixClass,RandomExponentialArraySeed-method}
\alias{RandomExponentialArraySeed}
\alias{DelayedArray,RandomExponentialArraySeed-method}
\alias{RandomExponentialArray}
\title{DelayedArray of chunked exponential values}
\usage{
RandomExponentialArraySeed(dim, rate = 1, chunkdim = NULL)

\S4method{DelayedArray}{RandomExponentialArraySeed}(seed)

RandomExponentialArray(dim, rate = 1, chunkdim = NULL)
}
\arguments{
\item{dim}{Integer vector of positive length, specifying the dimensions of the array.}

\item{rate}{Numeric vector used as \code{rate} in \code{\link{qexp}}.
Alternatively, a numeric array-like object with the same dimensions as \code{dim}.}

\item{chunkdim}{Integer vector of length equal to \code{dim}, containing the dimensions of each chunk.}

\item{seed}{A RandomExponentialArraySeed object.}
}
\value{
All constructors return an instance of a RandomExponentialArray object,
containing random draws from a exponential distribution with the specified parameters.
}
\description{
A \linkS4class{DelayedArray} subclass that performs on-the-fly sampling of exponentially distributed values.
}
\examples{
X <- RandomExponentialArraySeed(c(1e5, 1e5))
Y <- DelayedArray(X)
Y

# Fiddling with the distribution parameters:
X2 <- RandomExponentialArraySeed(c(1e5, 1e5), rate=runif(1e5))
Y2 <- DelayedArray(X2)
Y2

# Using another array as input:
library(Matrix)
rate <- rsparsematrix(1e5, 1e5, density=0.00001)
rate <- abs(DelayedArray(rate)) + 1
X3 <- RandomExponentialArraySeed(c(1e5, 1e5), rate=rate)
Y3 <- DelayedArray(X3)
Y3

}
\seealso{
The \linkS4class{RandomRandomArraySeed} class, for details on chunking and the distributional parameters.
}
\author{
Aaron Lun
}
