% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomNormalArray-class.R
\docType{class}
\name{RandomNormalArray-class}
\alias{RandomNormalArray-class}
\alias{RandomNormalArraySeed-class}
\alias{sampleDistrParam,RandomNormalArraySeed-method}
\alias{sampleDistrFun,RandomNormalArraySeed-method}
\alias{matrixClass,RandomNormalArraySeed-method}
\alias{RandomNormalArraySeed}
\alias{DelayedArray,RandomNormalArraySeed-method}
\alias{RandomNormalArray}
\title{DelayedArray of chunked normal values}
\usage{
RandomNormalArraySeed(dim, mean = 0, sd = 1, chunkdim = NULL)

\S4method{DelayedArray}{RandomNormalArraySeed}(seed)

RandomNormalArray(dim, mean = 0, sd = 1, chunkdim = NULL)
}
\arguments{
\item{dim}{Integer vector of positive length, specifying the dimensions of the array.}

\item{mean, sd}{Numeric vector used as \code{mean} and \code{sd}, respectively, in \code{\link{qnorm}}.
Alternatively, a numeric array-like object with the same dimensions as \code{dim}.}

\item{chunkdim}{Integer vector of length equal to \code{dim}, containing the dimensions of each chunk.}

\item{seed}{A RandomNormalArraySeed object.}
}
\value{
All constructors return an instance of a ChunkNormalArray object,
containing random draws from a normal distribution with the specified parameters.
}
\description{
A \linkS4class{DelayedArray} subclass that performs on-the-fly sampling of normally distributed values.
}
\examples{
X <- RandomNormalArraySeed(c(1e5, 1e5))
Y <- DelayedArray(X)
Y

# Fiddling with the distribution parameters:
X2 <- RandomNormalArraySeed(c(1e5, 1e5), mean=runif(1e5), sd=runif(1e5))
Y2 <- DelayedArray(X2)
Y2

# Using another array as input:
library(Matrix)
mean <- rsparsematrix(1e5, 1e5, density=0.00001)
X3 <- RandomNormalArraySeed(c(1e5, 1e5), mean=mean)
Y3 <- DelayedArray(X3)
Y3

}
\seealso{
The \linkS4class{RandomRandomArraySeed} class, for details on chunking and the distributional parameters.
}
\author{
Aaron Lun
}
