% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomPoissonArray-class.R
\docType{class}
\name{RandomPoissonArray-class}
\alias{RandomPoissonArray-class}
\alias{RandomPoissonArraySeed-class}
\alias{sampleDistrParam,RandomPoissonArraySeed-method}
\alias{sampleDistrFun,RandomPoissonArraySeed-method}
\alias{matrixClass,RandomPoissonArraySeed-method}
\alias{RandomPoissonArraySeed}
\alias{DelayedArray,RandomPoissonArraySeed-method}
\alias{RandomPoissonArray}
\title{DelayedArray of random Poisson values}
\usage{
RandomPoissonArraySeed(dim, lambda, chunkdim = NULL, sparse = FALSE)

\S4method{DelayedArray}{RandomPoissonArraySeed}(seed)

RandomPoissonArray(dim, lambda, chunkdim = NULL, sparse = FALSE)
}
\arguments{
\item{dim}{Integer vector of positive length, specifying the dimensions of the array.}

\item{lambda}{Numeric vector used as \code{lambda} in \code{\link{qpois}}.
Alternatively, a numeric array-like object with the same dimensions as \code{dim}.}

\item{chunkdim}{Integer vector of length equal to \code{dim}, containing the dimensions of each chunk.}

\item{sparse}{Logical scalar indicating whether the sampled matrix should be treated as sparse.}

\item{seed}{A RandomPoissonArraySeed object.}
}
\value{
All constructors return an instance of a RandomPoissonArray object,
containing random draws from a Poisson distribution with the specified parameters.
}
\description{
A \linkS4class{DelayedArray} subclass that performs on-the-fly sampling of Poisson-distributed values.
}
\examples{
X <- RandomPoissonArraySeed(c(1e5, 1e5), lambda=2)
Y <- DelayedArray(X)
Y

# Fiddling with the distribution parameters:
X2 <- RandomPoissonArraySeed(c(1e5, 1e5), lambda=runif(1e5), sparse=TRUE)
Y2 <- DelayedArray(X2)
Y2

# Using another array as input:
library(Matrix)
lambda <- rsparsematrix(1e5, 1e5, density=0.00001)
lambda <- abs(DelayedArray(lambda)) + 0.1
X3 <- RandomPoissonArraySeed(c(1e5, 1e5), lambda=lambda)
Y3 <- DelayedArray(X3)
Y3

}
\seealso{
The \linkS4class{RandomArraySeed} class, for details on chunking and the distributional parameters.
}
\author{
Aaron Lun
}
